inline ip_pools_fetch in the one call site
